Add product index, product details, missing img + index|category

+++ src/main/java/fr/poei/fines_saveurs_fo/service/ProductServiceImpl.java
@@ -21,6 +21,6 @@ public class ProductServiceImpl implements ProductService{
     }
 
     @Override
-    public Optional<Product> getById(Long id) { return productRepository.findById(id);}
+    public Optional<Product> getById(Long id) { return productRepository.findById(Math.toIntExact(id));}
 
 }
